+++ e2e/wdio/platform/tests/upload-collection.e2e-spec.ts
@@ -1,6 +1,7 @@
 import { UploadCollectionPo } from '../pages/upload-collection.po';
 
 import {
+    browserIsFirefox,
     click,
     getElementArrayLength,
     getElementPlaceholder,
@@ -11,6 +12,7 @@ import {
     scrollIntoView,
     sendKeys,
     setValue,
+    waitForElDisplayed,
     waitForNotDisplayed,
     waitForPresent
 } from '../../driver/wdio';
@@ -50,7 +52,8 @@ describe('Upload collection test suite', () => {
         listItem,
         moveButton,
         tableItem,
-        ghostButton
+        ghostButton,
+        dialog
     } = uploadCollectionPage;
 
     beforeAll(() => {
@@ -115,13 +118,16 @@ describe('Upload collection test suite', () => {
         checkClickabilityCancelButton(defaultExample);
     });
 
-    it('should check renaming folder', () => {
+    // skip due to https://github.com/SAP/fundamental-ngx/issues/7098
+    xit('should check renaming folder', () => {
         checkRenaming(turnOffExample);
         checkRenaming(defaultExample);
     });
 
-    // skipped due to broken layout https://github.com/SAP/fundamental-ngx/issues/6911
-    xit('should check moving folders', () => {
+    it('should check moving folders', () => {
+        if (browserIsFirefox()) {
+            return;
+        }
         checkMovingFolders(defaultExample);
         checkMovingFolders(turnOffExample);
     });
@@ -260,6 +266,7 @@ describe('Upload collection test suite', () => {
     function checkClickabilityCancelButton(selector: string): void {
         scrollIntoView(selector + transparentButton);
         click(defaultExample + transparentButton);
+        waitForElDisplayed(dialog);
         expect(isElementClickable(dialogCreateButton, 1)).toBe(true, '"Cancel" button not clickable');
         sendKeys(['Escape']);
     }