+++ libs/platform/src/lib/form/radio-group/radio/radio.component.spec.ts
@@ -75,28 +75,30 @@ describe('RadioButtonComponent', () => {
     });
 
     it('radio button with default property should be created', () => {
-        const inputElem = fixture.debugElement.query(By.css('input'));
+        const inputElem = <HTMLInputElement>fixture.debugElement.query(By.css('input')).nativeElement;
 
-        expect(inputElem.nativeElement.type).toEqual('radio');
-        expect(inputElem.nativeElement.getAttribute('id')).toBeTruthy();
-        expect(inputElem.nativeElement.getAttribute('ng-reflect-is-disabled')).toEqual('false');
-        expect(inputElem.nativeElement.getAttribute('ng-reflect-name')).toEqual('radio');
-        expect(inputElem.nativeElement.getAttribute('ng-reflect-value')).toEqual('1');
+        expect(inputElem.type).toEqual('radio');
+        expect(inputElem.getAttribute('id')).toBeTruthy();
+        expect(inputElem.getAttribute('ng-reflect-is-disabled')).toEqual('false');
+        expect(inputElem.getAttribute('ng-reflect-name')).toEqual('radio');
+        expect(inputElem.getAttribute('ng-reflect-value')).toEqual('1');
 
-        expect(inputElem.nativeElement.classList.contains('fd-radio')).toBeTruthy();
+        expect(inputElem.classList.contains('fd-radio')).toBeTruthy();
     });
 
     it('radio button should be compact, valid state and disable', () => {
         const inputElems = fixture.debugElement.queryAll(By.css('input'));
+        const inputElems1 = inputElems[1].nativeElement;
+        const inputElems2 = inputElems[2].nativeElement;
 
-        expect(inputElems[1].nativeElement.type).toEqual('radio');
-        expect(inputElems[1].nativeElement.getAttribute('id')).toBeTruthy();
-        expect(inputElems[2].nativeElement.getAttribute('ng-reflect-is-disabled')).toBeTruthy();
-        expect(inputElems[1].nativeElement.getAttribute('ng-reflect-name')).toEqual('radio');
-        expect(inputElems[1].nativeElement.getAttribute('ng-reflect-value')).toEqual('2');
+        expect(inputElems1.type).toEqual('radio');
+        expect(inputElems1.getAttribute('id')).toBeTruthy();
+        expect(inputElems2.getAttribute('ng-reflect-is-disabled')).toBeTruthy();
+        expect(inputElems1.getAttribute('ng-reflect-name')).toEqual('radio');
+        expect(inputElems1.getAttribute('ng-reflect-value')).toEqual('2');
 
-        expect(inputElems[1].nativeElement.classList.contains('fd-radio')).toBeTruthy();
-        expect(inputElems[1].nativeElement.classList.contains('fd-radio--compact')).toBeTruthy();
+        expect(inputElems1.classList.contains('fd-radio')).toBeTruthy();
+        expect(inputElems1.classList.contains('fd-radio--compact')).toBeTruthy();
     });
 
     it('radio click should should change control value', async () => {
@@ -104,11 +106,13 @@ describe('RadioButtonComponent', () => {
         fixture.detectChanges();
 
         const radioInputElems = fixture.debugElement.queryAll(By.css('input'));
+        const radioInputElems0 = radioInputElems[0].nativeElement;
+        const radioInputElems1 = radioInputElems[1].nativeElement;
 
         // first radio should be checked
         expect(component.selectedValue).toEqual(1);
-        expect(radioInputElems[0].nativeElement.getAttribute('aria-checked')).toEqual('true');
-        expect(radioInputElems[0].nativeElement.getAttribute('tabindex')).toEqual('0');
+        expect(radioInputElems0.getAttribute('aria-checked')).toEqual('true');
+        expect(radioInputElems0.getAttribute('tabindex')).toEqual('0');
 
         // click on second radio will check second radio and will change control value
         radioInputElems[1].nativeElement.click();
@@ -117,9 +121,9 @@ describe('RadioButtonComponent', () => {
         fixture.detectChanges();
         expect(component.selectedValue).toEqual(2);
 
-        expect(radioInputElems[0].nativeElement.getAttribute('aria-checked')).toEqual('false');
-        expect(radioInputElems[0].nativeElement.getAttribute('tabindex')).toEqual('-1');
-        expect(radioInputElems[1].nativeElement.getAttribute('aria-checked')).toEqual('true');
-        expect(radioInputElems[1].nativeElement.getAttribute('tabindex')).toEqual('0');
+        expect(radioInputElems0.getAttribute('aria-checked')).toEqual('false');
+        expect(radioInputElems0.getAttribute('tabindex')).toEqual('-1');
+        expect(radioInputElems1.getAttribute('aria-checked')).toEqual('true');
+        expect(radioInputElems1.getAttribute('tabindex')).toEqual('0');
     });
 });